Add build flag to disable packet injection
Packet may be a security concern. When it is not needed it is
better to no have it available at all.
This patch adds a build flag "nopacketinject" to disable packet
injection if needed.
