+++ api/server/no_packet_injector.go
@@ -1,0 +1,30 @@
+// +build !packetinject
+
+/*
+ * Copyright (C) 2016 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy ofthe License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specificlanguage governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package server
+
+import (
+	"fmt"
+
+	"github.com/skydive-project/skydive/api/types"
+)
+
+func (pi *PacketInjectorAPI) validateRequest(ppr *types.PacketInjection) error {
+	return fmt.Errorf("packet injection disabled because skydive was not built with \"packetinject\" build tag")
+}
