+++ api/server/packet_injector.go
@@ -1,5 +1,4 @@
-//go:generate sh -c "go run github.com/gomatic/renderizer --name=injection --resource=injectpacket --type=PacketInjection --title='Injection' --article=an swagger_operations.tmpl > packet_injector_swagger.go"
-//go:generate sh -c "go run github.com/gomatic/renderizer --name=injection --resource=injectpacket --type=PacketInjection --title='Injection' swagger_definitions.tmpl > packet_injector_swagger.json"
+// +build packetinject
 
 /*
  * Copyright (C) 2016 Red Hat, Inc.
@@ -26,42 +25,8 @@ import (
 	"net"
 
 	"github.com/skydive-project/skydive/api/types"
-	"github.com/skydive-project/skydive/graffiti/api/rest"
-	api "github.com/skydive-project/skydive/graffiti/api/server"
-	"github.com/skydive-project/skydive/graffiti/graph"
-	ge "github.com/skydive-project/skydive/gremlin/traversal"
-	shttp "github.com/skydive-project/skydive/http"
 	"github.com/skydive-project/skydive/topology"
 )
-
-type packetInjectorResourceHandler struct {
-	rest.ResourceHandler
-}
-
-// PacketInjectorAPI exposes the packet injector API
-type PacketInjectorAPI struct {
-	rest.BasicAPIHandler
-	Graph *graph.Graph
-}
-
-func (pirh *packetInjectorResourceHandler) Name() string {
-	return "injectpacket"
-}
-
-func (pirh *packetInjectorResourceHandler) New() rest.Resource {
-	return &types.PacketInjection{}
-}
-
-// Create allocates a new packet injection
-func (pi *PacketInjectorAPI) Create(r rest.Resource, opts *rest.CreateOptions) error {
-	ppr := r.(*types.PacketInjection)
-
-	if err := pi.validateRequest(ppr); err != nil {
-		return err
-	}
-	e := pi.BasicAPIHandler.Create(ppr, opts)
-	return e
-}
 
 func (pi *PacketInjectorAPI) validateRequest(ppr *types.PacketInjection) error {
 	pi.Graph.RLock()
@@ -121,37 +86,4 @@ func (pi *PacketInjectorAPI) validateRequest(ppr *types.PacketInjection) error {
 	}
 
 	return nil
-}
-
-func (pi *PacketInjectorAPI) getNode(gremlinQuery string) *graph.Node {
-	res, err := ge.TopologyGremlinQuery(pi.Graph, gremlinQuery)
-	if err != nil {
-		return nil
-	}
-
-	for _, value := range res.Values() {
-		switch value.(type) {
-		case *graph.Node:
-			return value.(*graph.Node)
-		default:
-			return nil
-		}
-	}
-	return nil
-}
-
-// RegisterPacketInjectorAPI registers a new packet injector resource in the API
-func RegisterPacketInjectorAPI(g *graph.Graph, apiServer *api.Server, authBackend shttp.AuthenticationBackend) (*PacketInjectorAPI, error) {
-	pia := &PacketInjectorAPI{
-		BasicAPIHandler: rest.BasicAPIHandler{
-			ResourceHandler: &packetInjectorResourceHandler{},
-			EtcdKeyAPI:      apiServer.EtcdKeyAPI,
-		},
-		Graph: g,
-	}
-	if err := apiServer.RegisterAPIHandler(pia, authBackend); err != nil {
-		return nil, err
-	}
-
-	return pia, nil
 }
